requirements.txt
mediapipe==0.10.33
opencv-python==4.11.0.86
scikit-learn==1.6.1
Pillow==11.3.0
PyYAML==6.0.3
numpy==2.4.4
ruff==0.11.4
